Change testes and implement upload

+++ src/main/webapp/resources/js/app/controllers.js
@@ -9,21 +9,37 @@ controllers.controller("MainController", ["$scope", "$rootScope" ,"$http", "$loc
 			alert("erro");
 		})
 	$scope.remover = function(id) {
-		$http.delete("/crud-veiculos/veiculos/delete/"+id+"").
-		success(function(data, status, headers, config) {
-			alert("Veiculo deletado");
-			//$location.path("/").replace();
-			window.location = "/crud-veiculos";
-		}).error(function(data, status, headers, config) {
-			alert("erro");
-		})
+		if(confirm("você tem certeza?")) {
+			$http.delete("/crud-veiculos/veiculos/delete/"+id+"").
+			success(function(data, status, headers, config) {
+				alert("Veiculo deletado");
+				// $location.path("/").replace();
+				window.location = "/crud-veiculos";
+			}).error(function(data, status, headers, config) {
+				alert("erro");
+			})
+		}
 	}
 }]);
 
 controllers.controller("VeiculoController", ["$scope", "$rootScope" ,"$http", "$location", function($scope, $rootScope ,$http, $location) {
+    $scope.model = {
+            fabricante: "",
+            modelo: "",
+            ano: "",
+            file:""
+        };
 	$scope.save = function() {
-		$http.post("/crud-veiculos/veiculos/save", $scope.veiculo, {
-		}).
+		$scope.model.fabricante = $scope.veiculo.fabricante;
+		$scope.model.modelo = $scope.veiculo.modelo;
+		$scope.model.ano = $scope.veiculo.ano;
+		$scope.model.file = $scope.veiculo.image;
+        $http({
+            method: 'POST',
+            url: "/crud-veiculos/veiculos/save",
+            //headers: { 'Content-Type': undefined },
+            data: $scope.model
+        }).
 			success(function(data, status, headers, config) {
 				alert("Veículo salvo com sucesso");
 				var path = "/edit/"+data.id;
@@ -35,7 +51,7 @@ controllers.controller("VeiculoController", ["$scope", "$rootScope" ,"$http", "$
 	
 }]);
 
-controllers.controller("EditVeiculoController", ["$scope", "$rootScope" ,"$http", "$location", "$routeParams", function($scope, $rootScope ,$http, $location, $routeParams, $window) {
+controllers.controller("EditVeiculoController", ["$scope", "$rootScope" ,"$http", "$location", "$routeParams", "$window", function($scope, $rootScope ,$http, $location, $routeParams, $window) {
 		$http.get("/crud-veiculos/veiculos/"+$routeParams.id+"").
 			success(function(data, status, headers, config) {
 				$rootScope.veiculo = data;
@@ -48,11 +64,12 @@ controllers.controller("EditVeiculoController", ["$scope", "$rootScope" ,"$http"
 				$http.post("/crud-veiculos/veiculos/update/"+$scope.veiculo.id+"", {
 					fabricante: $scope.veiculo.fabricante,
 					modelo: $scope.veiculo.modelo,
-					ano: $scope.veiculo.ano
+					ano: $scope.veiculo.ano,
+					file: $scope.veiculo.image
 				}).
 				success(function(data, status, headers, config) {
 					alert("Veículo alterado com sucesso");
-					$location.path("/").replace();
+					window.location = "/crud-veiculos"
 				}).error(function(data, status, headers, config) {
 					alert("erro");
 				})
@@ -60,7 +77,7 @@ controllers.controller("EditVeiculoController", ["$scope", "$rootScope" ,"$http"
 }]);
 
 controllers.controller("RemoveController", ["$scope", "$rootScope" ,"$http", "$location", "$routeParams", "$window",  function($scope, $rootScope ,$http, $location, $routeParams, $window) {
-	if($window.confirm("você tem certeza?")) {
+	if(confirm("você tem certeza?")) {
 		$http.delete("/crud-veiculos/delete/"+$routeParams.id+"").
 			success(function(data, status, headers, config) {
 				alert("Veiculo deletado");
